requirements.txt
absl-py==2.3.1
altair==5.5.0
numpy==2.3.0
tensorflow==2.20.0rc0
keras==3.11.1
opt_einsum==3.4.0
protobuf==5.29.4
termcolor==3.1.0
wrapt==1.17.2
opencv-python==4.11.0.86
pillow==11.2.1
streamlit==1.45.0